Isolate WMS GetPrint Outputs tests
